refactor: phrases hashmap index to usize

+++ src/encode.rs
@@ -21,8 +21,8 @@ pub fn run(text: &str) {
         // -1 is already filtered, gotta avoid filtering twice my god that would suck
         let mut taken = false;
         #[allow(clippy::cast_possible_truncation)]
-        for i in index..index + length as i32 - 1 {
-            if *ids.get(i as usize).unwrap() == -1 {
+        for i in index..index + length - 1 {
+            if *ids.get(i).unwrap() == -1 {
                 taken = true;
             }
         }
@@ -33,7 +33,7 @@ pub fn run(text: &str) {
     
             // replaces every character in the phrase with -1 besides the first character
             for i in 1..length {
-                let _ = std::mem::replace(&mut ids[index as usize + i], -1);
+                let _ = std::mem::replace(&mut ids[index + i], -1);
             }
         }
     }
